requirements.txt
streamlit
streamlit-mic-recorder
openai
python-dotenv